Introduces useCategoryContent talon
