+++ packages/peregrine/lib/talons/RootComponents/Category/index.js
@@ -1,0 +1,2 @@
+export { useCategoryContent } from './useCategoryContent';
+export { useNoProductsFound } from './useNoProductsFound';
